check for reservations locally
The modified reservations.py is still monitoring the MersenneForum thread but only writing to a local file. This gets transformed into a mock posting by res_post.php which is called by allseq.py instead of the MF postings.
Now we are not dependent on a MF moderator running the original reservations.py script.