Update yafu reserved seqs more often
It seems the priority system puts too much penalty on sequences that don't progress but are reserved. There are currently a lot of yafu reserved seqs with a high priority that have progressed in FDB but will not be updated until the 90 day limit is reached. This script runs twice a day to clear up the backlog for yafu.