add verify_merges.sh script and call it from updater.py
Script expects a file with lines that contain merging sequences. Each line must be of format "[0-9]*_[0-9]*(_[0-9]*)*" where the first sequence leader is the lowest in the line.
Currently sends an email when an error is encountered and when the merge is confirmed with the indices of the merge.

This also changes verify_terminations.sh and moves the verify_terminations option from aliqueit.ini to the commandline so the same ini file can be used for both verification of terminations and mergers.