refine merge verification
When adding a lot of new sequences it happens that the first update is not successful, when two seqs are not successful this was detected as a merge. The None check fixes that.
When a lot of merges are found the verification script takes very long and to prevent multiple instances from running it is now using a temp file and is also triggered when no merges are found but a non-empty merge file is found.